Renamed files, moved files. FIxed minor import problems in one or two controllers.
