+++ BE/src/Controllers/UploadInjestControllers.ts
@@ -7,7 +7,7 @@ import {
 } from "ResponseTypes";
 import { NextFunction } from "express";
 import PackageModel from "../Schemas/Package";
-import { CalculateStandaloneCost, CalculateTotalCost } from "../Services/Packages/Scoring/CalcPackageCost";
+import { CalculateStandaloneCost, CalculateTotalCost } from "../Services/Packages/Cost/CalcPackageCost";
 import {
     debloatUnzippedContent,
     debloatUploadedContent,
@@ -44,20 +44,34 @@ async function getGitHubDownload(repoURL: string): Promise<string> {
             }
             const repoData = await response.json();
             const defaultBranch = repoData.default_branch || "main"; // Fallback to 'main' if no default branch found
-            const defaultBranch = repoData.default_branch || "main"; // Fallback to 'main' if no default branch found
             return `https://github.com/${owner}/${repo}/archive/refs/heads/${defaultBranch}.zip`;
-        } catch (error) {
         } catch (error) {
             throw new Error(`Failed to fetch default branch: ${error}`);
         }
     }
-    throw new Error("Invalid GitHub URL");
     throw new Error("Invalid GitHub URL");
     //*/
 }
 
 const packagesDirectory = path.join(process.cwd(), "Data/Packages");
 const tempDirectory = packagesDirectory + "/.Temp";
+
+const findRepoUrl = (packageJson: any) => {
+    try {
+        if (packageJson.repository) {
+            if (typeof packageJson.repository === "string") {
+                return packageJson.repository; // Direct URL
+            }
+            if (typeof packageJson.repository === "object" && packageJson.repository.url) {
+                return packageJson.repository.url; // URL in object form
+            }
+        }
+        return null;
+    } catch (error) {
+        console.error(`Error reading or parsing package.json`);
+        return null;
+    }
+};
 
 async function cleanUp(tempID: string) {
     const tempFileZip = path.join(tempDirectory, tempID + ".zip");
@@ -74,7 +88,6 @@ export const UploadInjestController = asyncHandler(
     async (req: UploadInjestPackageRequest, res: UploadInjestNewPackageResponse, next: NextFunction) => {
         console.log("Entering Upload Process");
         const body = req.body;
-        const body = req.body;
         let repositoryUrl = body?.URL;
         let content = body?.Content;
         let binaryContent; // Meant to store the non-string encoded version
@@ -82,10 +95,7 @@ export const UploadInjestController = asyncHandler(
 
         const disqualifiedStandaloneSizeInGB = 750; // Approximately 1GB
         const disqualifiedTotalSizeInGB = 1000; // Approximately 1GB
-        const disqualifiedStandaloneSizeInGB = 750; // Approximately 1GB
-        const disqualifiedTotalSizeInGB = 1000; // Approximately 1GB
         const tempIDCeiling = 1000;
-        const tempID = Math.floor(Math.random() * tempIDCeiling + 1).toString();
         const tempID = Math.floor(Math.random() * tempIDCeiling + 1).toString();
         const tempFileZip = path.join(tempDirectory, tempID + ".zip");
         const tempUnzippedFileDirectory = path.join(tempDirectory, tempID);
@@ -97,53 +107,43 @@ export const UploadInjestController = asyncHandler(
             binaryContent = Buffer.from(base64Data, "base64");
             repositoryUrl = repositoryUrl as unknown as string; // Type casts it from "string | undefined" to "string"
         } else if (content == undefined && repositoryUrl != undefined) {
-        } else if (content == undefined && repositoryUrl != undefined) {
             // Confirmed that the repoURL exists, download content
             let repoDownloadURL: string;
-            try {
             try {
                 if (repositoryUrl.includes("github")) {
                     repoDownloadURL = await getGitHubDownload(repositoryUrl);
                 } else {
-                } else {
                     console.error("Not a valid URL");
                     responseMessage =
                         "There is missing field(s) in the PackageData or it is formed improperly (e.g. Content and URL are both set)";
-                    responseMessage =
-                        "There is missing field(s) in the PackageData or it is formed improperly (e.g. Content and URL are both set)";
+                    console.log(`/pacakge : ${responseMessage}`);
                     res.status(424).send(responseMessage);
                     return;
                 }
-            } catch (error) {
             } catch (error) {
                 console.log(error);
                 console.log("Error in getting the download link");
                 responseMessage =
                     "There is missing field(s) in the PackageData or it is formed improperly (e.g. Content and URL are both set)";
-                responseMessage =
-                    "There is missing field(s) in the PackageData or it is formed improperly (e.g. Content and URL are both set)";
+                console.log(
+                    `/pacakge : ${responseMessage} -> ${
+                        error instanceof Error ? error.message : "Unkown error in upload"
+                    }`
+                );
                 res.status(424).send(responseMessage);
                 return;
-            }
-            if (!fs.existsSync(tempDirectory)) {
-                // This is where data is downloaded before being examined.
-                fs.mkdirSync(tempDirectory);
             }
             console.log("Repo URL: " + repositoryUrl);
             console.log("Repo Download URL: " + repoDownloadURL);
             const response = await axios.get(repoDownloadURL, { responseType: "arraybuffer" });
             binaryContent = Buffer.from(response.data, "binary");
-            const response = await axios.get(repoDownloadURL, { responseType: "arraybuffer" });
-            binaryContent = Buffer.from(response.data, "binary");
             isExternal = true;
-        } else {
         } else {
             await cleanUp(tempID);
             console.error("Should only get here if both content and URL are undefined or they are defined");
             responseMessage =
                 "There is missing field(s) in the PackageData or it is formed improperly (e.g. Content and URL are both set)";
-            responseMessage =
-                "There is missing field(s) in the PackageData or it is formed improperly (e.g. Content and URL are both set)";
+            console.log(`/pacakge : ${responseMessage}`);
             res.status(424).send(responseMessage);
             return;
         }
@@ -151,20 +151,18 @@ export const UploadInjestController = asyncHandler(
         try {
             fs.promises.writeFile(tempFileZip, binaryContent);
         } catch (error) {
-        } catch (error) {
             await cleanUp(tempID);
             console.error(`Error: ${error}`);
             responseMessage =
                 "There is missing field(s) in the PackageData or it is formed improperly (e.g. Content and URL are both set)";
-            responseMessage =
-                "There is missing field(s) in the PackageData or it is formed improperly (e.g. Content and URL are both set)";
+            console.log(
+                `/pacakge : ${responseMessage} -> ${
+                    error instanceof Error ? error.message : "Unkown error in upload"
+                }`
+            );
             res.status(424).send(responseMessage);
             return;
         }
-
-        await unzipper.Open.buffer(binaryContent).then((directory) =>
-            directory.extract({ path: tempUnzippedFileDirectory })
-        );
 
         await unzipper.Open.buffer(binaryContent).then((directory) =>
             directory.extract({ path: tempUnzippedFileDirectory })
@@ -173,21 +171,12 @@ export const UploadInjestController = asyncHandler(
         const tempDirectoryListing = await fs.promises.readdir(tempUnzippedFileDirectory, {
             withFileTypes: true,
         });
-        const tempDirectoryListing = await fs.promises.readdir(tempUnzippedFileDirectory, {
-            withFileTypes: true,
-        });
         console.log("Directory listing length: " + tempDirectoryListing.length);
         if (tempDirectoryListing.length == 1) {
             nestedFolder = `/${tempDirectoryListing[0].name}`;
         }
         let packageJsonFile;
-        let packageJsonFile;
         try {
-            packageJsonFile = await fs.promises.readFile(
-                path.join(tempUnzippedFileDirectory + nestedFolder, "package.json"),
-                "utf-8"
-            );
-        } catch (error) {
             packageJsonFile = await fs.promises.readFile(
                 path.join(tempUnzippedFileDirectory + nestedFolder, "package.json"),
                 "utf-8"
@@ -197,28 +186,22 @@ export const UploadInjestController = asyncHandler(
             await cleanUp(tempID);
             responseMessage =
                 "There is missing field(s) in the PackageData or it is formed improperly (e.g. Content and URL are both set)";
-            responseMessage =
-                "There is missing field(s) in the PackageData or it is formed improperly (e.g. Content and URL are both set)";
+            console.log(
+                `/pacakge : ${responseMessage} -> ${
+                    error instanceof Error ? error.message : "Unkown error in upload"
+                }`
+            );
             res.status(424).send(responseMessage);
             return;
         }
-
 
         const packageJson = JSON.parse(packageJsonFile.toString());
         console.log("JSON Parsed");
         if (!isExternal) {
             try {
-                repositoryUrl = packageJson.repository.url;
-                if (repositoryUrl == undefined) {
-                    repositoryUrl = packageJson.homepage;
-                    if (repositoryUrl == undefined) {
-                        console.error("No Repo URL Found!");
-                        await cleanUp(tempID);
-                        responseMessage =
-                            "There is missing field(s) in the PackageData or it is formed improperly (e.g. Content and URL are both set)";
-                        res.status(424).send(responseMessage);
-                        return;
-                    }
+                repositoryUrl = findRepoUrl(packageJson);
+                if (!repositoryUrl) {
+                    throw new Error("Did not find repo url.");
                 }
             } catch (error) {
                 console.error(error);
@@ -226,8 +209,11 @@ export const UploadInjestController = asyncHandler(
                 await cleanUp(tempID);
                 responseMessage =
                     "There is missing field(s) in the PackageData or it is formed improperly (e.g. Content and URL are both set)";
-                responseMessage =
-                    "There is missing field(s) in the PackageData or it is formed improperly (e.g. Content and URL are both set)";
+                console.log(
+                    `/pacakge : ${responseMessage} -> ${
+                        error instanceof Error ? error.message : "Unkown error in upload"
+                    }`
+                );
                 res.status(424).send(responseMessage);
                 return;
             }
@@ -236,17 +222,15 @@ export const UploadInjestController = asyncHandler(
 
         // Checks if exists
         const queriedPackage = await PackageModel.exists({ repoUrl: repositoryUrl });
-        const queriedPackage = await PackageModel.exists({ repoUrl: repositoryUrl });
         let standaloneCost: number;
         let totalCost: number;
         if (queriedPackage !== null) {
             console.log(queriedPackage);
             await cleanUp(tempID);
             responseMessage = "Package exists already.";
+            console.log(`/pacakge : ${responseMessage}`);
             res.status(409).send(responseMessage);
             return;
-        } else {
-            // Checks if Disqualified
         } else {
             // Checks if Disqualified
             standaloneCost = await CalculateStandaloneCost(repositoryUrl); // No deps
@@ -254,6 +238,7 @@ export const UploadInjestController = asyncHandler(
             if (totalCost > disqualifiedTotalSizeInGB || standaloneCost > disqualifiedStandaloneSizeInGB) {
                 await cleanUp(tempID);
                 responseMessage = "Package is not uploaded due to disqualified rating.";
+                console.log(`/pacakge : ${responseMessage}`);
                 res.status(424).send(responseMessage);
             }
         }
@@ -284,23 +269,6 @@ export const UploadInjestController = asyncHandler(
                 standaloneCost,
                 totalCost
             )) + zipFileExtension;
-        const packageID =
-            (await buildMongoDBPackage(
-                {
-                    ...jsonRow,
-                    GoodPinningPracticeScore: 0,
-                    GoodPinningPracticeLatency: 0,
-                    PullRequestScore: 0,
-                    PullRequestLatency: 0,
-                },
-                repositoryUrl,
-                body.Name ? body.Name : packageJson.name ? packageJson.name : "Unknown", // Should never be unknown, but since this is a safety, it is here.
-                packageJson.version ? packageJson.version : "1.0.0",
-                packageJson.license ? packageJson.license : "Unknown",
-                isExternal,
-                standaloneCost,
-                totalCost
-            )) + zipFileExtension;
         if (body.debloat == true) {
             // Zip up, and store
             const isSuccessful = await debloatUnzippedContent(tempUnzippedFileDirectory);
@@ -311,15 +279,8 @@ export const UploadInjestController = asyncHandler(
                     path.join(packagesDirectory, packageID)
                 );
             } else {
-                zipContents(
-                    tempUnzippedFileDirectory,
-                    zipFileExtension,
-                    path.join(packagesDirectory, packageID)
-                );
-            } else {
                 await fs.promises.rename(tempFileZip, path.join(packagesDirectory, packageID));
             }
-        } else {
         } else {
             // Just move the existing zip to Data and rename to the ID.
             await fs.promises.rename(tempFileZip, path.join(packagesDirectory, packageID));
@@ -334,8 +295,8 @@ export const UploadInjestController = asyncHandler(
             //all fields are optional in data
             data: {},
         };
-        };
         await cleanupProcess;
+        console.log(`/pacakge : ${JSON.stringify(returnBody)}`);
         res.status(200).json(returnBody);
         return;
     }
