+++ BE/src/Providers/RepoEvaluator/ModuleEvaluator.ts
@@ -4,8 +4,9 @@ import { MetricName } from "./RepoComponents/Metrics_Scores/Metric.const";
 import { SubscoreCalculator } from "./RepoComponents/Metrics_Scores/ScoreCalculator";
 import { WeightSpec, FindWeightSpecByReceiver } from "./RepoComponents/Metrics_Scores/WeightSpec";
 import { EMPTY_WEIGHTSPEC, WeightSpecSet } from "./RepoComponents/Metrics_Scores/Weightspec.const";
-import { AsyncLoops, TryIndexOrDefaultTo } from "../../Utils/DSinc_Modules/DSinc_LoopsMaps";
+import { TryIndexOrDefaultTo } from "../../Utils/DSinc/Array";
 import { VersionDependence_Scorer, MergeRestriction_Scorer } from "./Functions/DSincScorers";
+import * as AsyncLoops from "../../Utils/DSinc/AsyncLoop";
 
 import {
     RAMPUP_WEIGHT_DEFAULT,
