+++ BE/src/Providers/RepoEvaluator/RepoComponents/Builders/Repository_Builder.ts
@@ -3,11 +3,8 @@ import { RepoID } from "../ID/RepoID";
 import { RepoID_Builder } from "../ID/RepoID_Builder";
 import { RepoURL } from "../URL/URLProcessor.interface";
 import { AsyncBuilder } from "../../Abstract_Builders";
-import { AsyncLoops } from "../../../../Utils/DSinc_Modules/DSinc_LoopsMaps";
-import {
-    IsType_RepoID,
-    IsType_RepoURL,
-} from "../../../../Utils/DSinc_Modules/CustomTypeGuards/ModEval_Guards";
+import * as AsyncLoops from "../../../../Utils/DSinc/AsyncLoop";
+import { IsType_RepoID, IsType_RepoURL } from "../../DevTools/ModEval_Guards";
 import { RepoScoreset } from "../Metrics_Scores/RepoScoreset";
 import { DEFAULT_WEIGHTS, WeightSpecSet } from "../Metrics_Scores/Weightspec.const";
