+++ BE/src/Providers/RepoEvaluator/RepoComponents/Metrics_Scores/WeightSpec.ts
@@ -1,4 +1,4 @@
-import * as DSincMath from "../../../../Utils/DSinc_Modules/DSinc_Math";
+import * as DSincMath from "../../../../Utils/DSinc/Math";
 import { MetricName } from "./Metric.const";
 
 export class WeightSpec {
