--- BE/src/Utils/CustomError.ts
@@ -1,14 +1,0 @@
-export type CustomErrorContent = {
-    message: string;
-    context?: { [key: string]: any };
-};
-
-export abstract class CustomError extends Error {
-    abstract readonly statusCode: number;
-    abstract readonly errors: CustomErrorContent[];
-    abstract readonly logging: boolean;
-    constructor(message: string) {
-        super(message);
-        Object.setPrototypeOf(this, CustomError.prototype);
-    }
-}
