+++ BE/src/Utils/DSinc/ErrorHandling.ts
@@ -89,3 +89,16 @@ export const ErrorWrapperForAsync = async <T>(
         process.exit(1);
     }
 };
+export abstract class CustomError extends Error {
+    abstract readonly statusCode: number;
+    abstract readonly errors: CustomErrorContent[];
+    abstract readonly logging: boolean;
+    constructor(message: string) {
+        super(message);
+        Object.setPrototypeOf(this, CustomError.prototype);
+    }
+}
+export type CustomErrorContent = {
+    message: string;
+    context?: { [key: string]: any };
+};
